Fix struct packing issue
